[Feat] 상품 삭제하기 기능 구현

+++ backend_mall/src/main/java/com/example/backend_mall/service/TodoServiceImpl.java
@@ -75,7 +75,7 @@ public class TodoServiceImpl implements TodoService{
         PageResponseDTO<TodoDTO> responseDTO = PageResponseDTO.<TodoDTO>withAll()
                 .dtoList(dtoList)
                 .pageRequestDTO(pageRequestDTO)
-                .totalCount(totalCount)
+                .totalCount((int)totalCount)
                 .build();
 
         return responseDTO;
